+++ zarrs/src/array.rs
@@ -1094,6 +1094,23 @@ mod tests {
         }
     }
 
+    #[test]
+    fn array_v2_none_c() {
+        array_v2_to_v3(
+            "tests/data/v2/array_none_C.zarr",
+            "tests/data/v3/array_none.zarr",
+        )
+    }
+
+    #[ignore] // FIXME: Reported upstream https://github.com/zarr-developers/zarr-python/issues/2675
+    #[test]
+    fn array_v2_none_f() {
+        array_v2_to_v3(
+            "tests/data/v2/array_none_F.zarr",
+            "tests/data/v3/array_none_tranpose.zarr",
+        )
+    }
+
     #[cfg(feature = "blosc")]
     #[test]
     #[cfg_attr(miri, ignore)]
@@ -1105,6 +1122,7 @@ mod tests {
     }
 
     #[cfg(feature = "blosc")]
+    #[ignore] // FIXME: Reported upstream https://github.com/zarr-developers/zarr-python/issues/2675
     #[test]
     #[cfg_attr(miri, ignore)]
     fn array_v2_blosc_f() {